Initial USDA Food Atlas Data in SAS tables
These tables provide a decent backdrop with data from 2006 through just before the pandemic.  This will be helpful to get a baseline so we can compare it to the more recent census data to determine communities most in need of investment as measured by food insecurity and/or other factors as we explore.